requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.0
arrow==1.2.3
async-timeout==4.0.2
attrs==23.1.0
bravado==10.4.1
bravado-core==5.17.1
certifi==2023.5.7
charset-normalizer==3.1.0
dateparser==1.1.8
fqdn==1.5.1
frozenlist==1.3.3
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
isoduration==20.11.0
jsonpointer==2.4
jsonref==1.1.0
jsonschema==4.17.3
monotonic==1.6
msgpack==1.0.5
msgpack-python==0.5.6
multidict==6.0.4
pybit==5.4.0
pycryptodome==3.18.0
pyrsistent==0.19.3
pyTelegramBotAPI==4.12.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-telegram-bot==20.3
pytz==2023.3
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
rfc3339-validator==0.1.4
rfc3987==1.3.8
simplejson==3.19.1
six==1.16.0
sniffio==1.3.0
swagger-spec-validator==3.0.3
typing_extensions==4.7.0
tzlocal==5.0.1
ujson==5.8.0
uri-template==1.3.0
urllib3==1.24.2
webcolors==1.13
websocket-client==1.6.1
websockets==11.0.3
yarl==1.9.2
